Added: Add setting to enable or disable background image loading
The user can enable or disable background image loading from the termux setting available in `Termux Style` preferences with this commit.
If some malformed image is stored as background, disable image loading in the setting.
Disabling the image from the setting will not delete image files and can be restored.